auth/requirements.txt
certifi==2017.11.5
chardet==3.0.4
click==6.7
coreapi==2.3.3
coreapi-cli==1.0.6
coreschema==0.0.4
defusedxml==0.5.0
Django==2.0
django-allauth==0.34.0
django-rest-auth==0.9.2
django-rest-swagger==2.0.7
djangorestframework==3.7.7
idna==2.6
itypes==1.1.0
Jinja2==2.10
Markdown==2.6.10
MarkupSafe==1.0
oauthlib==2.0.6
openapi-codec==1.3.2
psycopg2==2.7.3.2
Pygments==2.2.0
python3-openid==3.1.0
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
simplejson==3.13.2
six==1.9.0
uritemplate==3.0.0
urllib3==1.22
